Module stm32l4x6_hal::gpio

General Purpose Input / Output

This module provides common GPIO definitions that are available on all STM32L4x6 packages. By enabling features, you can use definitions for specific chips which may include additional GPIO lines. In that case, you will probably not want to use this module directly, but instead use it re-exported by the chip module.

Modules

stm32l476vg
GPIO specific to STM32L476VG

Structs

A
GPIO

AF0
Alternate function 0 (type state)

AF1
Alternate function 1 (type state)

AF2
Alternate function 2 (type state)

AF3
Alternate function 3 (type state)

AF4
Alternate function 4 (type state)

AF5
Alternate function 5 (type state)

AF6
Alternate function 6 (type state)

AF7
Alternate function 7 (type state)

AF8
Alternate function 8 (type state)

AF9
Alternate function 9 (type state)

AF10
Alternate function 10 (type state)

AF11
Alternate function 11 (type state)

AF12
Alternate function 12 (type state)

AF13
Alternate function 13 (type state)

AF14
Alternate function 14 (type state)

AF15
Alternate function 15 (type state)

AFRH
Opaque AFRH register

AFRL
Opaque AFRL register

B
GPIO

C
GPIO

Floating
Floating input (type state)

Input
Input mode (type state)

Led
Generic LED

MODER
Opaque MODER register

OTYPER
Opaque OTYPER register

OpenDrain
Open drain output (type state)

Output
Output mode (type state)

PA0
Specific Pin

PA1
Specific Pin

PA2
Specific Pin

PA3
Specific Pin

PA4
Specific Pin

PA5
Specific Pin

PA6
Specific Pin

PA7
Specific Pin

PA8
Specific Pin

PA9
Specific Pin

PA10
Specific Pin

PA11
Specific Pin

PA12
Specific Pin

PA13
Specific Pin

PA14
Specific Pin

PA15
Specific Pin

PB0
Specific Pin

PB1
Specific Pin

PB2
Specific Pin

PB3
Specific Pin

PB4
Specific Pin

PB5
Specific Pin

PB6
Specific Pin

PB7
Specific Pin

PB8
Specific Pin

PB9
Specific Pin

PB10
Specific Pin

PB11
Specific Pin

PB12
Specific Pin

PB13
Specific Pin

PB14
Specific Pin

PB15
Specific Pin

PC0
Specific Pin

PC1
Specific Pin

PC2
Specific Pin

PC3
Specific Pin

PC4
Specific Pin

PC5
Specific Pin

PC6
Specific Pin

PC7
Specific Pin

PC8
Specific Pin

PC9
Specific Pin

PC10
Specific Pin

PC11
Specific Pin

PC12
Specific Pin

PC13
Specific Pin

PC14
Specific Pin

PC15
Specific Pin

PUPDR
Opaque PUPDR register

PullDown
Pulled down input (type state)

PullUp
Pulled up input (type state)

PushPull
Push pull output (type state)

Traits

AltFun
Alternate Function Trait Implemented only for corresponding structs.

InputMode
Input Mode Trait Implemented only for corresponding structs.

OutputMode
Output Mode Trait Implemented only for corresponding structs.