Struct stm32l4x6_hal::gpio::stm32l496ag::gpio::PG13

pub struct PG13<MODE>(_);

Specific Pin

Methods

impl<MODE> PG13<MODE>

pub fn into_input<Mode: InputMode>(
self,
moder: &mut MODER<GPIOG>,
pupdr: &mut PUPDR<GPIOG>
) -> PG13<Input<Mode>>

Configures the PIN to operate as Input Pin according to Mode.

pub fn into_output<Mode: OutputMode>(
self,
moder: &mut MODER<GPIOG>,
otyper: &mut OTYPER<GPIOG>
) -> PG13<Output<Mode>>

Configures the PIN to operate as Output Pin according to Mode.

pub fn into_alt_fun<AF: AltFun>(
self,
moder: &mut MODER<GPIOG>,
afr: &mut AFRH<GPIOG>
) -> PG13<AF>

Configures the PIN to operate as Alternate Function.