Struct stm32l4x6_hal::gpio::PA6

pub struct PA6<MODE>(_);

Specific Pin

Methods

impl<MODE> PA6<MODE>

pub fn into_input<Mode: InputMode>(
self,
moder: &mut MODER<GPIOA>,
pupdr: &mut PUPDR<GPIOA>
) -> PA6<Input<Mode>>

Configures the PIN to operate as Input Pin according to Mode.

pub fn into_output<Mode: OutputMode>(
self,
moder: &mut MODER<GPIOA>,
otyper: &mut OTYPER<GPIOA>
) -> PA6<Output<Mode>>

Configures the PIN to operate as Output Pin according to Mode.

pub fn into_alt_fun<AF: AltFun>(
self,
moder: &mut MODER<GPIOA>,
afr: &mut AFRL<GPIOA>
) -> PA6<AF>

Configures the PIN to operate as Alternate Function.