Module stm32l4x6_hal::rcc::clocking

The clocking module contains representations of the various objects in the STM32L4x6 clock tree (see Reference Manual Figs 15 and 16) useful for wiring them up.

There are two main concepts: sources (enum variants) and clocks (structs). For example, the SYSCLK clock can be driven by any one of four sources: HSI, MSI, HSI, and PLLCLK. This knowledge is encoded in the SysClkSources enum.

Each enum variant contains information about the clock it represents. Some clocks are configurable, and thus have fields, and some are not. For example, the LSI (LowSpeedInternalRC) clock is always 32 kHz, but the MSI (MediumSpeedInternalRC) clock can be configured and thus has a frequency component.

To use them, compose them and feed them to, e.g., sysclk.

use stm32l4x6_hal::rcc::clocking;

let mut rcc = RCC.constrain();
let msi_clk = clocking::MediumSpeedInternalRC::new(8_000_000, false);
let sys_clk_src = clocking::SysClkSource::MSI(msi_clk);
let cfgr = rcc.cfgr.sysclk(sys_clk_src);

The PLL is a bit more complex because it is a source (PLLClkOutput) and also requires a source (PLLClkSource), but you compose the types similarly.

Structs

HighSpeedExternalOSC
High-speed external 4-48 MHz oscillator

HighSpeedInternal16RC
High-speed internal 16 MHz RC

MediumSpeedInternalRC
Medium-speed internal 100 kHz - 48 MHz RC

PLLClkOutput
PLLCLK output of PLL module

Enums

PLLClkSource
PLLADC2CLK output of PLLSAI2 Selectable PLL module input sources

RtcClkSource
Selectable input clocks to the RTC

SysClkSource
Selectable clocks for the SYSCLK signal (HCLK bus)

Traits

InputClock
Clocks (OSCs or RCs) that can be used as inputs to peripherals