Struct stm32l4x6_hal::serial::Serial

pub struct Serial<S, TX, RX, CK> {
    pub serial: S,
    // some fields omitted
}

Serial interface

Fields
serial: S
Methods

impl<UART: RawSerial, T: TX, R: RX> Serial<UART, T, R, DummyPin>

pub fn with_dummy(
serial: UART,
pins: (T, R),
baud_rate: u32,
clocks: &Clocks,
apb: &mut UART::APB
) -> Self

Initializes Serial with dummy CK

impl<UART: RawSerial, T: TX, R: RX, C: CK> Serial<UART, T, R, C>

pub fn new(
serial: UART,
pins: (T, R, C),
baud_rate: u32,
clocks: &Clocks,
apb: &mut UART::APB
) -> Self

Creates new instance of serial interface

Arguments:

serial - Serial interface.

pins - Pins used by serial.

baud_rate - Rate to set for TX and RX pins, See Reference Ch. 40.5.4 for details

apb - APBx corresponding to Serial.

It takes ownership of raw Serial object and corresponding PINs.

Pancis:

In debug mode the function checks if index of each PIN corresponds to Serial's index.

pub unsafe fn from_raw(serial: UART, pins: (T, R, C)) -> Self

Re-creates Serial instance from its components.

Note: it is up to user to ensure that Serial has been created using new previously

pub fn into_raw(self) -> (UART, (T, R, C))

Consumes self and returns Serial and PINS