Struct stm32l4x6_hal::spi::Spi

pub struct Spi<SPI, SCK, MISO, MOSI> { /* fields omitted */ }

SPI

Methods

impl<SPI: InnerSpi, S: SCK, MI: MISO, MO: MOSI> Spi<SPI, S, MI, MO>

pub fn new(
spi: SPI,
pins: (S, MI, MO),
freq: Hertz,
mode: Mode,
clocks: &Clocks,
apb: &mut SPI::APB
) -> Self

Creates new instance of SPI.

It takes ownership of raw SPI object and corresponding PINs.

Function performs following actions:

Reset and enable SPI;

Configure CR1;

Configure CR2;

Pancis:

In debug mode the function checks if index of each PIN corresponds to SPI's index.

pub unsafe fn from_raw(spi: SPI, pins: (S, MI, MO)) -> Self

Re-creates SPI instance from its components.

Note: it is up to user to ensure that SPI has been created using new previously

pub fn into_raw(self) -> (SPI, (S, MI, MO))

Consumes self and returns SPI and PINS