Struct stm32l4x6_hal::lcd::LCD

pub struct LCD { /* fields omitted */ }

LCD representations that provides access to HW LCD

Implements destructor that turns off LCD.

Methods

impl LCD

pub fn init_lse(
apb1: &mut APB1,
ahb: &mut AHB,
pwr: &mut Power,
bdcr: &mut BDCR
)

Initializes HW for LCD with LSE as clock source

Steps:

Enable peripheral clocks

Set LSE as RTC clock.

Turn on LCD's clock

pub fn new(lcd: LCD, config: Config) -> Self

Initializes LCD

Requirements

Pre-initialize HW using init method.

Configure LCD GPIO pins as alternative functions. You should check board's documentation for pins.

Verify configuration through validate function

Steps:

Turns off.

Reset RAM registers and set update request.

Performs configuration.

Turns on.

pub fn validate(
lcd: &mut LCD,
bdcr: &mut BDCR,
configuration: &Config
) -> ValidationResult

Performs validation of settings.

HSE clock is not supported yet...

pub fn is_enabled(&mut self) -> bool

Returns whether LCD is enabled or not

pub fn is_ready(&mut self) -> bool

Returns whether LCD is ready or not

pub fn configure(&mut self, config: Config)

Performs LCD's configuration

pub fn reset_ram(&mut self)

Resets LCD's RAM.

To have effect user must request update.

pub fn update_request(&mut self)

Requests to transfer written data to buffer by setting SR's UDR bit

Note: Once set, it can be cleared only by hardware In addition to that until value is cleared, RAM is write-protected.

No update can occur until display shall be enabled.

pub fn on(&mut self)

Turns LCD on by setting CR's EN bit

pub fn off(&mut self)

Turns LCD off by clearing CR's EN bit

pub fn subscribe(&mut self, event: Event)

Starts listening for an event

pub fn unsubscribe(&mut self, event: Event)

Stops listening for an event

pub fn write_ram<I: Index>(&mut self, data: u32)

Writes into RAM by index.

pub fn into_raw(self) -> LCD